# src/auslib/web/admin/swagger/api.yml
swagger: '2.0'
info:
  version: '1.0'
  title: "Balrog's Admin API"
  termsOfService: "https://wiki.mozilla.org/Balrog"
  license:
    name: "Mozilla Public License 2.0"
    url: "https://www.mozilla.org/en-US/MPL/2.0/"
  contact:
    name: "Mozilla Release Engineering"
    url: "https://github.com/mozilla-releng/balrog"
    email: "release-engineering@lists.mozilla.org"
  description: >
    **Balrog** is the software that runs the server side component of the update
    system used by Firefox and other Mozilla products. It is the successor to *AUS*
    (Application Update Service), which did not scale to our current needs nor allow
    us to adapt to more recent business requirements. Balrog helps us ship updates faster
    and with much more flexibility than we’ve had in the past.

# host and scheme can be omitted for a more dynamic association.
# In this case, the host and scheme used to serve the API documentation
# will be used for API calls. For example, if Swagger UI-based documentation is
# hosted at https://api.example.com/apidocs/index.html, “try it out” API calls
# will be directed to https://api.example.com.

#schemes:
#  - http
#  - https
# basePath: "/api"
# host: "localhost:8080"
# Swagger-UI bug : fix TODO week 13
...
parameters:
...
  changedByParam:
    name: changed_by
    in: query
    description: The username who made the requested changes.
    type: string
    format: ascii
    x-nullable: true
    required: false
  changeIdParam:
    name: change_id
    in: query
    description: The change_id value to filter hsitory by.
    type: integer
    x-nullable: true
    required: false
  historyDataVersionParam:
    name: data_version
    in: query
    description: data_version field value to filter history by.
    type: integer
    x-nullable: true
    required: false
...
  timestampFromParam:
    name: timestamp_from
    in: query
    description: timestamp from/above which to filter history.
    type: integer
    x-nullable: true
    required: false
  timestampToParam:
    name: timestamp_to
    in: query
    description: timestamp to/below which history is filtered.
    type: integer
    x-nullable: true
    required: false
...
paths:
...
  /permissions/history:
    get:
      summary: Returns a list of permissions and it's scheduled change history.
      description: >
        Returns a list of permissions and it's scheduled change history.
      tags:
        - "History Permissions"
      operationId: auslib.web.common.history_all.permissions
      consumes: []
      produces: [application/json]
      externalDocs:
        url: "http://mozilla-balrog.readthedocs.io/en/latest/admin_api.html#put"
        description: "Returns a list of permissions and permissions scheduled change history"
      parameters:
        - $ref: '#/parameters/productParam'
        - $ref: '#/parameters/limitParam'
        - $ref: '#/parameters/pageParam'
        - $ref: '#/parameters/changedByParam'
        - $ref: '#/parameters/changeIdParam'
        - $ref: '#/parameters/historyDataVersionParam'
        - $ref: '#/parameters/timestampFromParam'
        - $ref: '#/parameters/timestampToParam'
      responses:
        '200':
          description: Successfully fetched permissions and it's scheduled change history
        '400':
          $ref: '#/responses/invalidFormData'
...